BD/seed/requirements.txt
faker==18.0.0
psycopg2-binary==2.9.3